mgr/cephadm: moving cert_key_store class to cert_mgr module
