+++ src/pybind/mgr/cephadm/serve.py
@@ -141,8 +141,8 @@ class CephadmServe:
         for d in self.mgr.cache.get_daemons_by_type('grafana'):
             host = d.hostname
             assert host is not None
-            cert = self.mgr.cert_key_store.get_cert('grafana_cert', host=host)
-            key = self.mgr.cert_key_store.get_key('grafana_key', host=host)
+            cert = self.mgr.cert_mgr.get_cert('grafana_cert', host=host)
+            key = self.mgr.cert_mgr.get_key('grafana_key', host=host)
             if (not cert or not cert.strip()) and (not key or not key.strip()):
                 # certificate/key are empty... nothing to check
                 return
